requirements.txt
Pillow==3.3.1
selenium==2.53.6
numpy==1.11.2
tensorflow==0.11.0rc0

# rewardfunction only
scipy==0.18.1
scikit-image==0.12.3
matplotlib==1.5.3
scikit-learn==0.18

# Mac OS X, CPU only, Python 2.7:
https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-0.11.0rc0-py2-none-any.whl

ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0

# OpenCV:
# brew tap homebrew/science
# brew install opencv
# Symlink to virtualenv
# ln -s /usr/local/lib/python2.7/site-packages/cv2.so usr/local/lib/python2.7/site-packages/cv.py .venv/lib/python2.7/site-packages